refactor: organize components into feature folders
