--- src/components/index.ts
@@ -1,7 +1,0 @@
-export { default as CourseEditor } from './CourseEditor';
-export { default as CourseSelector } from './CourseSelector';
-export { default as PlayerSetup } from './PlayerSetup';
-export { default as ScoreCard } from './ScoreCard';
-export { default as PlayerIcon } from './PlayerIcon';
-export { default as PlayerSelect } from './PlayerSelect';
-export { default as PlayerHeader } from './PlayerHeader';
