+++ src/features/index.ts
@@ -1,0 +1,7 @@
+export { default as CourseEditor } from './course/CourseEditor';
+export { default as CourseSelector } from './course/CourseSelector';
+export { default as PlayerSetup } from './player/PlayerSetup';
+export { default as ScoreCard } from './score/ScoreCard';
+export { default as PlayerIcon } from './player/PlayerIcon';
+export { default as PlayerSelect } from './player/PlayerSelect';
+export { default as PlayerHeader } from './player/PlayerHeader';
