+++ src/features/player/PlayerSelect.tsx
@@ -1,5 +1,5 @@
 import { useState, useRef, useEffect } from 'react';
-import { Player } from '../types/golf';
+import { Player } from '../../types/golf';
 import PlayerIcon from './PlayerIcon';
 
 interface PlayerSelectProps {
